refactor: move commons
