+++ src/styles/common/cover.css
