+++ src/types/notion/blocks/paragraph.ts
@@ -1,11 +1,11 @@
 import type { BlockBlockObject, BlockBlockProps } from "./block.js";
-import type { RichTextItem } from "../richText/richTextItem.js";
+import type { RichText } from "../richText/richText.js";
 import type { ParagraphBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints.js";
 import type { ComponentType } from "react";
 
 export type ParagraphBlockObject = ParagraphBlockObjectResponse & {
   paragraph: {
-    rich_text: Array<RichTextItem>;
+    rich_text: RichText;
     children?: Array<BlockBlockObject>;
   };
 };
